+++ Ruddarr/Models/Movies/MovieLookup.swift
@@ -108,6 +108,14 @@ class MovieLookup {
         return queries[query] ?? nil
     }
 
+    func updateItem(_ movie: Movie) {
+        if let index = items?.firstIndex(where: { $0.tmdbId == movie.tmdbId }) {
+            items?[index] = movie
+        }
+
+        queries["tmdb:\(movie.tmdbId)"] = movie
+    }
+
     // consider caching this for performance
     var sortedItems: [Movie] {
         let items = items ?? []